requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
tweepy==4.14.0
google-generativeai==0.3.2
python-multipart==0.0.6
openai==1.3.0
agents==0.1.0